test: enable parallel mode for wal_off tests
- `memtx_memory` is increased, because the test `lua` after `tuple`
failed with the error:
`Failed to allocate 368569 bytes in slab allocator for memtx_tuple`
despite `collectgarbage('collect')` calls after cases with huge/many
tuples.
The statistics before the allocation fail gives the following values:
```
box.slab.info()
---
- items_size: 72786472
  items_used_ratio: 4.43%
  quota_size: 107374592
  quota_used_ratio: 93.75%
  arena_used_ratio: 6.1%
  items_used: 3222376
  quota_used: 100663296
  arena_size: 100663296
  arena_used: 6105960
```
The reason of the fail seems to be a slab memory fragmentation.  It is
not clear for now whether we should consider this as a tarantool issue.

- `snapshot_stress` test counts snapshot files present in the
working directory and can reach the default 'checkpoint_count' value
`2` if a previous test write its snapshots before.

- restarting the default server w/o cleaning a working directory
can leave a snapshot that holds a state saved at the middle of a test,
before dropping of the space 'tweedledum' (because WAL is disabled),
that can cause the error `Space 'tweedledum' already exists` for a
following test.

Part of #2436