test: increase timeout to check replica status
The replica status is checked 100 times, each check within
`replica_timeout`. Refactor the code to get properly upstream.
Then in loop with little sleep check upstreams status until
it is not in follow mode. If count of checks is more than 200
break the loop with error. The value 200 and little sleep 0.001
choosed suitably to `replica_timeout` and `replica_connect_timeout`.

Part of #2436, #3232